fix(agents): break providers <-> discovery import cycle (CodeQL)
CodeQL flagged the function-scope imports between providers.py and
discovery.py as a cyclic import. The cycle is real even if Python
tolerates it at runtime: discovery needs the canonical
name -> provider mapping; providers needs the strict command
classifier that discovery owns.

Extract the small, dependency-free pieces (the KNOWN_PROVIDERS
frozenset, the cursor-family map, and provider_from_classified_name)
into a new app.agents.provider_ids module. Both providers.py and
discovery.py now depend on provider_ids only, so providers.py can
import classify_command_provider from discovery at module load
without re-creating the cycle.

providers.py re-exports KNOWN_PROVIDERS and
provider_from_classified_name so existing call sites (tests, the
sampler, the meter / token-source registries) keep working without
changes.
